自动打分

+++ src/main/java/com/hw/mapper/RadioStreamResultTabMapper.java
@@ -17,7 +17,7 @@ public interface RadioStreamResultTabMapper {
 
     public Page selectList(RadioStreamResultTabQuery query);
 
-    public RadioStreamResultTab selectById(Integer id);
+    public RadioStreamResultTab selectById(Long id);
 
     public Integer delete(RadioStreamResultTab bean);
